recommender-engine/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
python-dotenv==0.17.0
psycopg2-binary==2.8.6
pandas==1.2.3